requirements.txt
asgiref==3.7.2
Django==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.4
